change file, preconditions and bodies for MARC bib tests
